+++ docs/CHANGELOG.md
@@ -9,9 +9,15 @@ Changed in NOW
 - `Added` SerialWriter.writeBytes(OutputStream out, byte[], int, int).
 - `Added` Wad.getContent(int, int) : byte[]. Implementing classes changed as well.
 - `Added` Palette mixing functions plus copying.
+- `Added` WadFile.addAllData[At](...) "No Flush" variants.
+- `Added` GraphicUtils.setColormap(...).
+- `Added` GraphicUtils.createColormapsFromGraphic(...).
 - `Fixed` GraphicUtils.createPicture(...) / GraphicUtils.createFlat(...) handles non-opaque pixels accurately.
 - `Changed` Some test refactoring.
 - `Changed` WadBuffer - made some protected fields private and merged the header in with content.
+- `Changed` Flat and Picture now extend IndexedGraphic.
+- `Changed` WadUtils.cleanEntries(..) shortened.
+- `Changed` WadUtils.extract(..) made more efficient.
 
 
 Changed in 2.1.1